merge, 충돌해결
